requirements.txt
numpy
opencv-python
imutils
matplotlib
basenet
pytesseract
argparse
tensorflow
keras
keras-applications
gTTS
playsound
swtloc
Shapely
pillow
pandas
tqdm